Rename pkg/controller to pkg/operator
